refactor(claw-server-rust): split audit and session tab repositories
